Update differentialMethylation.R
I just tried to use the limmaP function as an isolated function, so I might have misunderstood the usage of the function, but: 
if I understand it correctly, you reorder the methylation matrix putting samples of group1 and then samples of group2. 
However the adtjustment.table was not ordered. Or is it ordered already somewhere else?